Refactor lessons list screen layout
